Make 'partial' fetching dependant on test paramater.
Before: There is a hardcoded limit on the range of repos whose
shas arefetched for the dictionary. This is for testing purposes.

After: This limit now depends on the setting of the environment
variable [$loki_test_max]. If this is set to '5' only the first
five repos will have their shas fetched.
for the dictionary.